StatusID became Long, added method getRequests

+++ src/main/java/ru/rsatu/pojo/Orders.java
@@ -14,16 +14,16 @@ public class Orders extends PanacheEntity {
     private Long clientID;
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.EAGER) // order из файла OrdersDetails
     public List<OrdersDetails> orderDetails;
-    private Integer statusID;
+    private Long statusID;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private Date creationDate;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private Date lastUpdateOn;
 
-	public Integer getStatusID() {
+	public Long getStatusID() {
 		return statusID;
 	}
-	public void setStatusID(Integer i) {
+	public void setStatusID(Long i) {
 		this.statusID = i;
 	}
 	public Long getClientID() {
